main code
This code use reinforcement learning and specifically the Q-learning approach for training a unicycle robot to make decision-based on a state-action matrix. Actions are the direction of robot. learning will converge when the robot achieved the goal. I'd be more than grateful if you could comment and make my code more complete.